Added state check to prevent readonly check after object is removed
git-svn-id: https://svn.mssm.nl/sandbox/ontopia-jdo@707 869db993-6338-4487-9605-f03c488beb64

+++ src/main/java/net/ontopia/topicmaps/impl/jdo/TMObject.java
@@ -25,6 +25,7 @@ import java.util.HashSet;
 import java.util.Set;
 import javax.jdo.JDOException;
 import javax.jdo.JDOHelper;
+import javax.jdo.ObjectState;
 import javax.jdo.PersistenceManager;
 import javax.jdo.Query;
 import javax.jdo.annotations.IdGeneratorStrategy;
@@ -119,6 +120,8 @@ public abstract class TMObject implements TMObjectIF {
 
 	@Override
 	public void remove() {
+		ObjectState state = JDOHelper.getObjectState(this);
+		if ((state == ObjectState.PERSISTENT_DELETED) || (state == ObjectState.PERSISTENT_NEW_DELETED)) return;
 		if (isReadOnly()) throw new ReadOnlyException();
 		beforeRemove();
 		getPersistenceManager().deletePersistent(this);
